# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lightbearer Audio Broadcasting Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface BroadcasterProfile {
  id: number;
  name: string;
  username: string;
  email: string;
  phone?: string | null;
  bio?: string | null;
  avatarUrl?: string | null;
  coverUrl?: string | null;
  followerCount: number;
  broadcastCount: number;
  isLive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  broadcaster: BroadcasterProfile;
  token?: string;
}

export interface UpdateBroadcasterBody {
  name?: string;
  bio?: string;
  phone?: string;
  avatarUrl?: string;
  coverUrl?: string;
}

export interface Broadcast {
  id: number;
  broadcasterId: number;
  broadcaster?: BroadcasterProfile;
  title: string;
  description?: string | null;
  thumbnailUrl?: string | null;
  venue?: string | null;
  minister?: string | null;
  tags: string[];
  isLive: boolean;
  listenerCount: number;
  recordingUrl?: string | null;
  isRecorded: boolean;
  savedToDraft: boolean;
  startedAt: string;
  endedAt?: string | null;
}

export interface CreateBroadcastBody {
  broadcasterId: number;
  title: string;
  description?: string;
  thumbnailUrl?: string;
  venue?: string;
  minister?: string;
  /**
   * @minItems 5
   * @maxItems 15
   */
  tags: string[];
  isRecorded?: boolean;
}

export interface UpdateBroadcastBody {
  isLive?: boolean;
  listenerCount?: number;
  endedAt?: string;
  recordingUrl?: string;
  savedToDraft?: boolean;
}

export interface Recording {
  id: number;
  broadcastId?: number | null;
  broadcasterId: number;
  title: string;
  url: string;
  thumbnailUrl?: string | null;
  durationSeconds?: number | null;
  isPublic: boolean;
  isDraft: boolean;
  createdAt: string;
}

export interface CreateRecordingBody {
  broadcastId?: number;
  broadcasterId: number;
  title: string;
  url: string;
  thumbnailUrl?: string;
  durationSeconds?: number;
  isPublic: boolean;
  isDraft: boolean;
}

export interface PlatformStats {
  liveBroadcasts: number;
  totalListeners: number;
  totalBroadcasters: number;
  totalRecordings: number;
}

export type ListBroadcastsParams = {
  live?: boolean;
};

export type SearchBroadcastsParams = {
  q: string;
};