# lib/api-zod/src/generated/types/broadcast.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lightbearer Audio Broadcasting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BroadcasterProfile } from "./broadcasterProfile";

export interface Broadcast {
  id: number;
  broadcasterId: number;
  broadcaster?: BroadcasterProfile;
  title: string;
  description?: string | null;
  thumbnailUrl?: string | null;
  venue?: string | null;
  minister?: string | null;
  tags: string[];
  isLive: boolean;
  listenerCount: number;
  recordingUrl?: string | null;
  isRecorded: boolean;
  savedToDraft: boolean;
  startedAt: Date;
  endedAt?: Date | null;
}