# lib/api-zod/src/generated/types/createBroadcastBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lightbearer Audio Broadcasting Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBroadcastBody {
  broadcasterId: number;
  title: string;
  description?: string;
  thumbnailUrl?: string;
  venue?: string;
  minister?: string;
  /**
   * @minItems 5
   * @maxItems 15
   */
  tags: string[];
  isRecorded?: boolean;
}