# lib/api-zod/src/generated/types/createRecordingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lightbearer Audio Broadcasting Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRecordingBody {
  broadcastId?: number;
  broadcasterId: number;
  title: string;
  url: string;
  thumbnailUrl?: string;
  durationSeconds?: number;
  isPublic: boolean;
  isDraft: boolean;
}