# lib/api-zod/src/generated/types/updateBroadcasterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lightbearer Audio Broadcasting Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBroadcasterBody {
  name?: string;
  bio?: string;
  phone?: string;
  avatarUrl?: string;
  coverUrl?: string;
}